fix utils logger and tracer
