+++ cli/src/commands/prove.rs
@@ -1,8 +1,8 @@
 use anstyle::*;
 use anyhow::Result;
 use clap::Parser;
-use sp1_prover::io::SP1Stdin;
-use sp1_sdk::utils::{setup_logger, setup_tracer};
+use sp1_prover::utils::{setup_logger, setup_tracer};
+use sp1_prover::SP1Stdin;
 use sp1_sdk::ProverClient;
 use std::fs;
 use std::time::Instant;
