+++ core/src/syscall/precompiles/keccak256/air.rs
@@ -146,10 +146,9 @@ where
 
 #[cfg(test)]
 mod test {
-    use crate::io::SP1Stdin;
+    use crate::io::{SP1PublicValues, SP1Stdin};
     use crate::runtime::Program;
     use crate::stark::{RiscvAir, StarkGenericConfig};
-    use crate::utils::Buffer;
     use crate::utils::{run_and_prove, setup_logger, tests::KECCAK256_ELF, BabyBearPoseidon2};
 
     use rand::Rng;
@@ -186,7 +185,7 @@ mod test {
 
         let program = Program::from(KECCAK256_ELF);
         let (proof, public_values) = run_and_prove(program, &stdin.buffer, config);
-        let mut public_values = Buffer::from(&public_values);
+        let mut public_values = SP1PublicValues::from(&public_values);
 
         let config = BabyBearPoseidon2::new();
         let mut challenger = config.challenger();
