+++ eval/src/main.rs
@@ -6,8 +6,8 @@ use csv::WriterBuilder;
 use serde::Serialize;
 use sp1_core::runtime::{Program, Runtime};
 use sp1_core::utils::{get_cycles, prove_core, BabyBearBlake3, BabyBearKeccak, BabyBearPoseidon2};
-use sp1_prover::io::{SP1ProofWithIO, SP1PublicValues, SP1Stdin};
 use sp1_prover::SP1ProverImpl;
+use sp1_prover::{SP1ProofWithIO, SP1PublicValues, SP1Stdin};
 use std::fmt;
 use std::fs::OpenOptions;
 use std::io;
